Move draft post components to their own folder
