+++ frontend/src/components/draftPost/DraftPost.tsx
@@ -3,8 +3,8 @@ import { Post } from '@/sharedDataModel';
 import { Profile, CoStream, CoID } from 'cojson';
 import { Resolved, useJazz } from 'jazz-react';
 import { useCallback, useState } from 'react';
-import { Button } from './ui/button';
-import { Textarea } from './ui/textarea';
+import { Button } from '../ui/button';
+import { Textarea } from '../ui/textarea';
 import { cn } from '@/lib/utils';
 import { DraftPostImage } from './DraftPostImage';
 import { DraftPostScheduler } from './DraftPostScheduler';
