+++ frontend/src/components/draftPost/DraftPostScheduler.tsx
@@ -1,6 +1,6 @@
 import { Post } from '@/sharedDataModel';
-import { Button } from './ui/button';
-import { Input } from './ui/input';
+import { Button } from '../ui/button';
+import { Input } from '../ui/input';
 import { toDateTimeLocal } from '@/lib/dates';
 import { Resolved } from 'jazz-react';
